Refactoring PhylogenyController
Moved stuff to NewickAlgorithm. The PhylogenyController had too much
responsibility.
