+++ src/main/java/parsers/NewickTreeParser.java
@@ -9,7 +9,7 @@ import java.nio.charset.Charset;
 import java.util.Stack;
 import java.util.StringTokenizer;
 
-import newick.NewickTree;
+import gui.phylogeny.model.NewickTree;
 
 /**
  *
